Hardcoded for now automatic transactions
We read the notification stream from the device for payment
notifications sent from Google Wallet and attempt to parse them into
automatic transactions.

It's not perfect.

One - we need to redo the DB table.
Instead of creating the transaction as a DbTransaction directly we
should create a DbAutomaticNotification table with entries:

sbn_id, sbn_key, transaction_id (fk)

This way we can reliably avoid duplications if we see the same
notification with the same key. The transaction_id is (fk) to the DAN
table so that if a user removes the auto transaction from their list we
wipe the entry from the DAN table too

DAN created transactions will have category of at least [ automatic ]
which will be a system level category

Categories need to have an additional field for level with values

"user" | "system"